restructuring into component
