+++ index.css
@@ -30,10 +30,11 @@
   padding: 5px;
   width: 200px;
   background-color: rgb(234, 231, 231);
+  box-shadow: 0px 4px 6px rgb(0, 0, 0, 0.1);
 }
 
 .restaurant-card:hover {
-  border: 1px solid black;
+  border: 1px solid transparent;
   cursor: pointer;
 }
