+++ src/component/api/exampleData.js
@@ -1,0 +1,1652 @@
+const restaurantData = [
+  {
+    info: {
+      id: "31192",
+      name: "Sri Udupi Park - Opp Bescom",
+      cloudinaryImageId: "qhiy6coolxjequzdstn0",
+      locality: "Phoenix Marketcity Mall",
+      areaName: "Mahadevapura",
+      costForTwo: "₹200 for two",
+      cuisines: ["South Indian", "North Indian", "Jain", "Chinese", "Desserts"],
+      avgRating: 4.2,
+      veg: true,
+      feeDetails: {
+        restaurantId: "31192",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 4700,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 4700,
+      },
+      parentId: "381856",
+      avgRatingString: "4.2",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 27,
+        lastMileTravel: 5,
+        serviceability: "SERVICEABLE",
+        slaString: "27 mins",
+        lastMileTravelString: "5.0 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-07 22:01:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      aggregatedDiscountInfoV2: {},
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/sri-udupi-park-opp-bescom-phoenix-marketcity-mall-mahadevapura-bangalore-31192",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "156663",
+      name: "Sri Lakshmi Vaibhava",
+      cloudinaryImageId: "hxdoygnqwwi6i6fenhjo",
+      locality: "C.T. Street",
+      areaName: "Marathahalli",
+      costForTwo: "₹200 for two",
+      cuisines: ["South Indian", "North Indian"],
+      avgRating: 4.3,
+      veg: true,
+      feeDetails: {
+        restaurantId: "156663",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 2900,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 2900,
+      },
+      parentId: "14682",
+      avgRatingString: "4.3",
+      totalRatingsString: "5K+",
+      sla: {
+        deliveryTime: 21,
+        lastMileTravel: 1.4,
+        serviceability: "SERVICEABLE",
+        slaString: "21 mins",
+        lastMileTravelString: "1.4 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-07 22:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      aggregatedDiscountInfoV2: {},
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/sri-lakshmi-vaibhava-c-t-street-marathahalli-bangalore-156663",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "235721",
+      name: "Chaayos Chai+Snacks=Relax",
+      cloudinaryImageId: "2beebd3688e63e027ee8a11c1c8a72c8",
+      locality: "Kundalahalli",
+      areaName: "Marathahalli",
+      costForTwo: "₹250 for two",
+      cuisines: [
+        "Bakery",
+        "Beverages",
+        "Chaat",
+        "Desserts",
+        "Home Food",
+        "Italian",
+        "Maharashtrian",
+        "Snacks",
+        "Street Food",
+        "Sweets",
+      ],
+      avgRating: 4.3,
+      feeDetails: {
+        restaurantId: "235721",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 3500,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 3500,
+      },
+      parentId: "281782",
+      avgRatingString: "4.3",
+      totalRatingsString: "5K+",
+      sla: {
+        deliveryTime: 24,
+        lastMileTravel: 3,
+        serviceability: "SERVICEABLE",
+        slaString: "24 mins",
+        lastMileTravelString: "3.0 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-08 04:00:00",
+        opened: true,
+      },
+      badges: {
+        textExtendedBadges: [
+          {
+            iconId: "guiltfree/GF_Logo_android_3x",
+            shortDescription: "options available",
+            fontColor: "#7E808C",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "",
+                  fontColor: "#7E808C",
+                  iconId: "guiltfree/GF_Logo_android_3x",
+                  shortDescription: "options available",
+                },
+              },
+            ],
+          },
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "60% OFF",
+        subHeader: "UPTO ₹120",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/chaayos-chai-snacks-relax-kundalahalli-marathahalli-bangalore-235721",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "5167",
+      name: "Theobroma",
+      cloudinaryImageId: "b033728dcb0101ceb19bff0e1e1f6474",
+      locality: "2nd stage",
+      areaName: "Indiranagar",
+      costForTwo: "₹800 for two",
+      cuisines: ["Bakery", "Desserts", "Beverages"],
+      avgRating: 4.6,
+      feeDetails: {
+        restaurantId: "5167",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 7900,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 7900,
+      },
+      parentId: "1040",
+      avgRatingString: "4.6",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 32,
+        lastMileTravel: 7.4,
+        serviceability: "SERVICEABLE",
+        slaString: "32 mins",
+        lastMileTravelString: "7.4 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-07 23:45:00",
+        opened: true,
+      },
+      badges: {
+        imageBadges: [
+          {
+            imageId: "newg.png",
+            description: "Gourmet",
+          },
+        ],
+      },
+      isOpen: true,
+      aggregatedDiscountInfoV2: {},
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "Gourmet",
+                  imageId: "newg.png",
+                },
+              },
+            ],
+          },
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/theobroma-2nd-stage-indiranagar-bangalore-5167",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "667",
+      name: "Chai Point",
+      cloudinaryImageId: "qjhkcmjv598tcxkvh6ln",
+      locality: "Yamalur",
+      areaName: "Marathahalli",
+      costForTwo: "₹150 for two",
+      cuisines: [
+        "Bakery",
+        "Beverages",
+        "Maharashtrian",
+        "Snacks",
+        "Street Food",
+        "South Indian",
+        "Punjabi",
+        "Chaat",
+        "Indian",
+        "American",
+        "North Indian",
+        "Fast Food",
+        "Desserts",
+        "Cafe",
+        "Healthy Food",
+        "Home Food",
+      ],
+      avgRating: 4.3,
+      feeDetails: {
+        restaurantId: "667",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 2900,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 2900,
+      },
+      parentId: "1607",
+      avgRatingString: "4.3",
+      totalRatingsString: "1K+",
+      sla: {
+        deliveryTime: 26,
+        lastMileTravel: 2.5,
+        serviceability: "SERVICEABLE",
+        slaString: "26 mins",
+        lastMileTravelString: "2.5 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-07 22:00:00",
+        opened: true,
+      },
+      badges: {
+        textExtendedBadges: [
+          {
+            iconId: "guiltfree/GF_Logo_android_3x",
+            shortDescription: "options available",
+            fontColor: "#7E808C",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "",
+                  fontColor: "#7E808C",
+                  iconId: "guiltfree/GF_Logo_android_3x",
+                  shortDescription: "options available",
+                },
+              },
+            ],
+          },
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "60% OFF",
+        subHeader: "UPTO ₹120",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/chai-point-yamalur-marathahalli-bangalore-667",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "32177",
+      name: "Sri Udupi Park",
+      cloudinaryImageId: "i4zll6pl8zwxuq6lpgpl",
+      locality: "AECS Layout",
+      areaName: "Marathahalli",
+      costForTwo: "₹150 for two",
+      cuisines: [
+        "South Indian",
+        "North Indian",
+        "Chaat",
+        "Beverages",
+        "Chinese",
+      ],
+      avgRating: 4,
+      veg: true,
+      feeDetails: {
+        restaurantId: "32177",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 2900,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 2900,
+      },
+      parentId: "4284",
+      avgRatingString: "4.0",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 24,
+        lastMileTravel: 2,
+        serviceability: "SERVICEABLE",
+        slaString: "24 mins",
+        lastMileTravelString: "2.0 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-07 21:31:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "10% OFF",
+        subHeader: "UPTO ₹40",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/sri-udupi-park-aecs-layout-marathahalli-bangalore-32177",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "997",
+      name: "A2B - Adyar Ananda Bhavan",
+      cloudinaryImageId: "wcqvllaujsmtz1ji2vlp",
+      locality: "Munekolala",
+      areaName: "Marathahalli",
+      costForTwo: "₹300 for two",
+      cuisines: ["South Indian", "North Indian", "Sweets", "Chinese"],
+      avgRating: 4.3,
+      veg: true,
+      feeDetails: {
+        restaurantId: "997",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 3500,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 3500,
+      },
+      parentId: "22",
+      avgRatingString: "4.3",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 22,
+        lastMileTravel: 3,
+        serviceability: "SERVICEABLE",
+        slaString: "22 mins",
+        lastMileTravelString: "3.0 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-07 22:30:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "20% OFF",
+        subHeader: "UPTO ₹50",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/a2b-adyar-ananda-bhavan-munekolala-marathahalli-bangalore-997",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "40668",
+      name: "BIB - Breakfast in the Box",
+      cloudinaryImageId: "t9hhgu5jps4qcwum6gux",
+      locality: "Raj Palaya",
+      areaName: "Hoodi",
+      costForTwo: "₹350 for two",
+      cuisines: ["American", "Continental"],
+      avgRating: 3.9,
+      feeDetails: {
+        restaurantId: "40668",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 4700,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 4700,
+      },
+      parentId: "45396",
+      avgRatingString: "3.9",
+      totalRatingsString: "5K+",
+      sla: {
+        deliveryTime: 33,
+        lastMileTravel: 5.9,
+        serviceability: "SERVICEABLE",
+        slaString: "33 mins",
+        lastMileTravelString: "5.9 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-07 15:00:00",
+        opened: true,
+      },
+      badges: {
+        textExtendedBadges: [
+          {
+            iconId: "guiltfree/GF_Logo_android_3x",
+            shortDescription: "options available",
+            fontColor: "#7E808C",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "",
+                  fontColor: "#7E808C",
+                  iconId: "guiltfree/GF_Logo_android_3x",
+                  shortDescription: "options available",
+                },
+              },
+            ],
+          },
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "20% OFF",
+        subHeader: "UPTO ₹50",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/bib-breakfast-in-the-box-raj-palaya-hoodi-bangalore-40668",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "59067",
+      name: "Samosa Singh",
+      cloudinaryImageId: "77baefd8a5e319c828b4d7dff7260644",
+      locality: "Munekolalla",
+      areaName: "Marathahalli",
+      costForTwo: "₹150 for two",
+      cuisines: ["Snacks", "North Indian", "Beverages"],
+      avgRating: 4.1,
+      veg: true,
+      feeDetails: {
+        restaurantId: "59067",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 2900,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 2900,
+      },
+      parentId: "5639",
+      avgRatingString: "4.1",
+      totalRatingsString: "5K+",
+      sla: {
+        deliveryTime: 29,
+        lastMileTravel: 1.9,
+        serviceability: "SERVICEABLE",
+        slaString: "29 mins",
+        lastMileTravelString: "1.9 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-14 00:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "10% OFF",
+        subHeader: "UPTO ₹40",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/samosa-singh-munekolalla-marathahalli-bangalore-59067",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "6911",
+      name: "FreshMenu",
+      cloudinaryImageId: "6c4de7b481d3614370e5062b2f33e7d6",
+      locality: "Gulmohar road",
+      areaName: "Marathahalli",
+      costForTwo: "₹250 for two",
+      cuisines: [
+        "Continental",
+        "Chinese",
+        "Oriental",
+        "Asian",
+        "Healthy Food",
+        "Fast Food",
+        "Indian",
+        "Desserts",
+      ],
+      avgRating: 3.9,
+      feeDetails: {
+        restaurantId: "6911",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 2900,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 2900,
+      },
+      parentId: "398",
+      avgRatingString: "3.9",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 29,
+        lastMileTravel: 1.2,
+        serviceability: "SERVICEABLE",
+        slaString: "29 mins",
+        lastMileTravelString: "1.2 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-08 00:00:00",
+        opened: true,
+      },
+      badges: {
+        textExtendedBadges: [
+          {
+            iconId: "guiltfree/GF_Logo_android_3x",
+            shortDescription: "options available",
+            fontColor: "#7E808C",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "",
+                  fontColor: "#7E808C",
+                  iconId: "guiltfree/GF_Logo_android_3x",
+                  shortDescription: "options available",
+                },
+              },
+            ],
+          },
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "EVERY ITEM",
+        subHeader: "@ ₹99",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/freshmenu-gulmohar-road-marathahalli-bangalore-6911",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "149682",
+      name: "Samosa Party",
+      cloudinaryImageId: "thzhbierhb2qxhqsl0bo",
+      locality: "Munnekollal",
+      areaName: "Marathahalli",
+      costForTwo: "₹100 for two",
+      cuisines: [
+        "Fast Food",
+        "Snacks",
+        "Beverages",
+        "Chaat",
+        "North Indian",
+        "Street Food",
+        "Sweets",
+        "Desserts",
+        "Punjabi",
+        "Bakery",
+      ],
+      avgRating: 4.3,
+      feeDetails: {
+        restaurantId: "149682",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 2900,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 2900,
+      },
+      parentId: "6364",
+      avgRatingString: "4.3",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 16,
+        lastMileTravel: 1.1,
+        serviceability: "SERVICEABLE",
+        slaString: "16 mins",
+        lastMileTravelString: "1.1 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-07 22:00:00",
+        opened: true,
+      },
+      badges: {},
+      select: true,
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "60% OFF",
+        subHeader: "UPTO ₹120",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/samosa-party-munnekollal-marathahalli-bangalore-149682",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "490585",
+      name: "NOTO - Healthy Ice Cream",
+      cloudinaryImageId: "544f9e6dce1efb7054e4ac5d98aba049",
+      locality: "Munnekolala",
+      areaName: "Marathalli",
+      costForTwo: "₹250 for two",
+      cuisines: ["Ice Cream"],
+      avgRating: 4.3,
+      veg: true,
+      feeDetails: {
+        restaurantId: "490585",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 2900,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 2900,
+      },
+      parentId: "7158",
+      avgRatingString: "4.3",
+      totalRatingsString: "500+",
+      sla: {
+        deliveryTime: 18,
+        lastMileTravel: 1.6,
+        serviceability: "SERVICEABLE",
+        slaString: "18 mins",
+        lastMileTravelString: "1.6 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-08 05:45:00",
+        opened: true,
+      },
+      badges: {
+        textExtendedBadges: [
+          {
+            iconId: "guiltfree/GF_Logo_android_3x",
+            shortDescription: "brand",
+            fontColor: "#7E808C",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "",
+                  fontColor: "#7E808C",
+                  iconId: "guiltfree/GF_Logo_android_3x",
+                  shortDescription: "brand",
+                },
+              },
+            ],
+          },
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹125 OFF",
+        subHeader: "ABOVE ₹499",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/noto-healthy-ice-cream-munnekolala-marathalli-bangalore-490585",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "103789",
+      name: "Starbucks Coffee",
+      cloudinaryImageId: "4df3497f1460dfd1ecd8125222f6d362",
+      locality: "Ramagondanahalli",
+      areaName: "Whitefield",
+      costForTwo: "₹400 for two",
+      cuisines: [
+        "Beverages",
+        "Cafe",
+        "Snacks",
+        "Desserts",
+        "Bakery",
+        "Ice Cream",
+      ],
+      avgRating: 4.4,
+      feeDetails: {
+        restaurantId: "103789",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 3500,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 3500,
+      },
+      parentId: "195515",
+      avgRatingString: "4.4",
+      totalRatingsString: "1K+",
+      sla: {
+        deliveryTime: 25,
+        lastMileTravel: 3,
+        serviceability: "SERVICEABLE",
+        slaString: "25 mins",
+        lastMileTravelString: "3.0 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-07 23:59:00",
+        opened: true,
+      },
+      badges: {
+        textExtendedBadges: [
+          {
+            iconId: "guiltfree/GF_Logo_android_3x",
+            shortDescription: "options available",
+            fontColor: "#7E808C",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "",
+                  fontColor: "#7E808C",
+                  iconId: "guiltfree/GF_Logo_android_3x",
+                  shortDescription: "options available",
+                },
+              },
+            ],
+          },
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "10% OFF",
+        subHeader: "ABOVE ₹800",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/starbucks-coffee-ramagondanahalli-whitefield-bangalore-103789",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "43850",
+      name: "The Nosh House",
+      cloudinaryImageId: "t6c2esjgibs3nluwubvk",
+      locality: "Kasavanahalli",
+      areaName: "Sarjapur Road",
+      costForTwo: "₹600 for two",
+      cuisines: [
+        "Continental",
+        "Indian",
+        "Desserts",
+        "Italian",
+        "American",
+        "Mediterranean",
+        "Mexican",
+        "Thai",
+        "Asian",
+        "Pan-Asian",
+        "North Indian",
+        "Home Food",
+        "Healthy Food",
+        "Snacks",
+        "Pizzas",
+        "Pastas",
+        "Salads",
+        "Beverages",
+        "Bakery",
+      ],
+      avgRating: 4.5,
+      feeDetails: {
+        restaurantId: "43850",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 10900,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 10900,
+      },
+      parentId: "21005",
+      avgRatingString: "4.5",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 49,
+        lastMileTravel: 11.7,
+        serviceability: "SERVICEABLE",
+        slaString: "49 mins",
+        lastMileTravelString: "11.7 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-07 22:30:00",
+        opened: true,
+      },
+      badges: {
+        textExtendedBadges: [
+          {
+            iconId: "guiltfree/GF_Logo_android_3x",
+            shortDescription: "options available",
+            fontColor: "#7E808C",
+          },
+        ],
+      },
+      isOpen: true,
+      aggregatedDiscountInfoV2: {},
+      type: "C",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "",
+                  fontColor: "#7E808C",
+                  iconId: "guiltfree/GF_Logo_android_3x",
+                  shortDescription: "options available",
+                },
+              },
+            ],
+          },
+        },
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/the-nosh-house-kasavanahalli-sarjapur-road-bangalore-43850",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "9420",
+      name: "Madurai Idly Shop",
+      cloudinaryImageId: "d0caf075899a1a86a2f4d96ea8dbb668",
+      locality: "Munekollal",
+      areaName: "Marathahalli",
+      costForTwo: "₹150 for two",
+      cuisines: ["South Indian"],
+      avgRating: 3.9,
+      veg: true,
+      feeDetails: {
+        restaurantId: "9420",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 2900,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 2900,
+      },
+      parentId: "613",
+      avgRatingString: "3.9",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 17,
+        lastMileTravel: 0.7,
+        serviceability: "SERVICEABLE",
+        slaString: "17 mins",
+        lastMileTravelString: "0.7 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-07 23:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "30% OFF",
+        subHeader: "UPTO ₹75",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/madurai-idly-shop-munekollal-marathahalli-bangalore-9420",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "72821",
+      name: "Burger King",
+      cloudinaryImageId: "e33e1d3ba7d6b2bb0d45e1001b731fcf",
+      locality: "Dodda Kannali",
+      areaName: "Kaikondrahalli",
+      costForTwo: "₹350 for two",
+      cuisines: ["Burgers", "American"],
+      avgRating: 4.2,
+      feeDetails: {
+        restaurantId: "72821",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 7400,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 7400,
+      },
+      parentId: "166",
+      avgRatingString: "4.2",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 42,
+        lastMileTravel: 8.6,
+        serviceability: "SERVICEABLE",
+        slaString: "42 mins",
+        lastMileTravelString: "8.6 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-08 03:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "EVERY ITEM",
+        subHeader: "@ ₹129",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/burger-king-dodda-kannali-kaikondrahalli-bangalore-72821",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "4750",
+      name: "Subway",
+      cloudinaryImageId: "1ace5fa65eff3e1223feb696c956b38b",
+      locality: "Kaikondrahalli",
+      areaName: "Sarjapur Road",
+      costForTwo: "₹350 for two",
+      cuisines: ["Salads", "Snacks", "Desserts", "Beverages"],
+      avgRating: 4.1,
+      feeDetails: {
+        restaurantId: "4750",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 5900,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 5900,
+      },
+      parentId: "2",
+      avgRatingString: "4.1",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 40,
+        lastMileTravel: 8,
+        serviceability: "SERVICEABLE",
+        slaString: "40 mins",
+        lastMileTravelString: "8.0 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-08 03:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      aggregatedDiscountInfoV2: {},
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/subway-kaikondrahalli-sarjapur-road-bangalore-4750",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "129157",
+      name: "Sri Alok Park",
+      cloudinaryImageId: "d1qegfnppziz8rurhthx",
+      locality: "ITPL Main Road",
+      areaName: "Whitefield",
+      costForTwo: "₹200 for two",
+      cuisines: [
+        "South Indian",
+        "North Indian",
+        "Snacks",
+        "Desserts",
+        "Juices",
+      ],
+      avgRating: 4.2,
+      veg: true,
+      feeDetails: {
+        restaurantId: "129157",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 5900,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 5900,
+      },
+      parentId: "373323",
+      avgRatingString: "4.2",
+      totalRatingsString: "5K+",
+      sla: {
+        deliveryTime: 34,
+        lastMileTravel: 7.6,
+        serviceability: "SERVICEABLE",
+        slaString: "34 mins",
+        lastMileTravelString: "7.6 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-07 23:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "10% OFF",
+        subHeader: "UPTO ₹40",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/sri-alok-park-itpl-main-road-whitefield-bangalore-129157",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "7069",
+      name: "Cafe Coffee Day",
+      cloudinaryImageId: "b925459fbc1faf59f02f7289eb079a74",
+      locality: "2nd stage",
+      areaName: "Indiranagar",
+      costForTwo: "₹300 for two",
+      cuisines: [
+        "Beverages",
+        "Cafe",
+        "Snacks",
+        "Desserts",
+        "Burgers",
+        "Ice Cream",
+        "Bakery",
+        "Fast Food",
+      ],
+      avgRating: 4.2,
+      feeDetails: {
+        restaurantId: "7069",
+        fees: [
+          {
+            name: "BASE_DISTANCE",
+            fee: 7400,
+          },
+          {
+            name: "BASE_TIME",
+          },
+          {
+            name: "ANCILLARY_SURGE_FEE",
+          },
+        ],
+        totalFee: 7400,
+      },
+      parentId: "1",
+      avgRatingString: "4.2",
+      totalRatingsString: "5K+",
+      sla: {
+        deliveryTime: 42,
+        lastMileTravel: 8.2,
+        serviceability: "SERVICEABLE",
+        slaString: "42 mins",
+        lastMileTravelString: "8.2 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2023-09-14 00:00:00",
+        opened: true,
+      },
+      badges: {
+        textExtendedBadges: [
+          {
+            iconId: "guiltfree/GF_Logo_android_3x",
+            shortDescription: "options available",
+            fontColor: "#7E808C",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "",
+                  fontColor: "#7E808C",
+                  iconId: "guiltfree/GF_Logo_android_3x",
+                  shortDescription: "options available",
+                },
+              },
+            ],
+          },
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹150 OFF",
+        subHeader: "ABOVE ₹699",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/cafe-coffee-day-2nd-stage-indiranagar-bangalore-7069",
+      type: "WEBLINK",
+    },
+  },
+];
+
+export default restaurantData;
